ASCII/pyproject.toml
[project]
name = "ascii"
version = "0.1.0"
description = ""
authors = [
    {name = "Hesam Pourabbasian",email = "hesampourabbasian6@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "ascii", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
